Task2/DjangoProject/requirements.txt
alpha-vantage==2.1.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
djangorestframework==3.10.3
idna==2.8
numpy==1.17.3
pandas==0.25.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
